[Security Solution] render graph preview in new document flyout in Security Solution and Discover
