+++ x-pack/solutions/security/plugins/security_solution/public/flyout/document_details/left/tabs/visualize_tab.test.tsx
@@ -7,7 +7,7 @@
 
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import { useGraphPreview } from '../../shared/hooks/use_graph_preview';
+import { useGraphPreview } from '../../../../flyout_v2/document/hooks/use_graph_preview';
 import { useUpsellingComponent } from '../../../../common/hooks/use_upselling';
 import { useExpandableFlyoutState } from '@kbn/expandable-flyout';
 import { useDocumentDetailsContext } from '../../shared/context';
@@ -21,7 +21,7 @@ const mockAnalyzeGraphTestId = 'analyze-graph';
 const mockSessionViewTestId = 'session-view';
 
 // Mock all required dependencies
-jest.mock('../../shared/hooks/use_graph_preview');
+jest.mock('../../../../flyout_v2/document/hooks/use_graph_preview');
 jest.mock('../../../../common/hooks/use_upselling');
 jest.mock('@kbn/expandable-flyout');
 jest.mock('../../shared/context');
@@ -80,9 +80,7 @@ describe('VisualizeTab', () => {
     });
 
     (useDocumentDetailsContext as jest.Mock).mockReturnValue({
-      getFieldsData: jest.fn(),
-      dataAsNestedObject: {},
-      dataFormattedForFieldBrowser: {},
+      searchHit: { _id: 'test-id', _index: 'test-index', fields: {} },
     });
   });
