+++ x-pack/solutions/security/plugins/security_solution/public/flyout/document_details/right/components/visualizations_section.tsx
@@ -13,11 +13,12 @@ import { useExpandSection } from '../../../../flyout_v2/shared/hooks/use_expand_
 import { AnalyzerPreviewContainer } from '../../../../flyout_v2/document/components/analyzer_preview_container';
 import { SessionPreviewContainer } from '../../../../flyout_v2/document/components/session_preview_container';
 import { ExpandableSection } from '../../../../flyout_v2/shared/components/expandable_section';
-import { GraphPreviewContainer } from './graph_preview_container';
+import { GraphPreviewContainer } from '../../../../flyout_v2/document/components/graph_preview_container';
 import { useDocumentDetailsContext } from '../../shared/context';
-import { useGraphPreview } from '../../shared/hooks/use_graph_preview';
+import { useGraphPreview } from '../../../../flyout_v2/document/hooks/use_graph_preview';
 import { useNavigateToAnalyzer } from '../../shared/hooks/use_navigate_to_analyzer';
 import { useNavigateToSessionView } from '../../shared/hooks/use_navigate_to_session_view';
+import { useNavigateToGraphVisualization } from '../../shared/hooks/use_navigate_to_graph_visualization';
 import {
   VISUALIZATION_SECTION_TEST_ID,
   VISUALIZATION_SECTION_TITLE,
@@ -34,26 +35,13 @@ export const VisualizationsSection = memo(() => {
     title: KEY,
     defaultValue: false,
   });
-  const {
-    dataAsNestedObject,
-    getFieldsData,
-    dataFormattedForFieldBrowser,
-    isRulePreview,
-    eventId,
-    indexName,
-    scopeId,
-    isPreviewMode,
-    searchHit,
-  } = useDocumentDetailsContext();
+  const { isRulePreview, eventId, indexName, scopeId, isPreviewMode, searchHit } =
+    useDocumentDetailsContext();
 
   const hit = useMemo(() => buildDataTableRecord(searchHit as EsHitRecord), [searchHit]);
 
   // Decide whether to show the graph preview or not
-  const { hasGraphData } = useGraphPreview({
-    getFieldsData,
-    ecsData: dataAsNestedObject,
-    dataFormattedForFieldBrowser,
-  });
+  const { hasGraphData } = useGraphPreview({ hit });
 
   const { navigateToAnalyzer } = useNavigateToAnalyzer({
     eventId,
@@ -68,6 +56,12 @@ export const VisualizationsSection = memo(() => {
     isFlyoutOpen: true,
     scopeId,
     isPreviewMode,
+  });
+  const { navigateToGraphVisualization } = useNavigateToGraphVisualization({
+    eventId,
+    indexName,
+    isFlyoutOpen: true,
+    scopeId,
   });
 
   return (
@@ -95,7 +89,12 @@ export const VisualizationsSection = memo(() => {
       {hasGraphData && (
         <>
           <EuiSpacer />
-          <GraphPreviewContainer />
+          <GraphPreviewContainer
+            hit={hit}
+            onShowGraph={navigateToGraphVisualization}
+            showIcon={!isPreviewMode}
+            disableNavigation={isRulePreview || isPreviewMode}
+          />
         </>
       )}
     </ExpandableSection>
