+++ x-pack/solutions/security/plugins/security_solution/public/flyout_v2/document/components/graph_preview.tsx
