+++ x-pack/solutions/security/plugins/security_solution/public/flyout_v2/document/components/visualizations_section.test.tsx
@@ -62,6 +62,14 @@ jest.mock('./session_preview_container', () => ({
   SessionPreviewContainer: () => <div data-test-subj="sessionPreviewContainerMock" />,
 }));
 
+jest.mock('./graph_preview_container', () => ({
+  GraphPreviewContainer: () => <div data-test-subj="graphPreviewContainerMock" />,
+}));
+
+jest.mock('../hooks/use_graph_preview', () => ({
+  useGraphPreview: jest.fn(() => ({ hasGraphData: true })),
+}));
+
 const createMockHit = (flattened: DataTableRecord['flattened']): DataTableRecord =>
   ({
     id: '1',
@@ -146,5 +154,6 @@ describe('VisualizationsSection', () => {
     expect(
       getByTestId(EXPANDABLE_PANEL_CONTENT_TEST_ID(ANALYZER_PREVIEW_TEST_ID))
     ).toBeInTheDocument();
+    expect(getByTestId('graphPreviewContainerMock')).toBeInTheDocument();
   });
 });
